fix: resolve published package import issues with CLI-as-main-entry approach
- CLI package now serves as main entry point routing to MCP server or CLI
- Implement dual import strategy: workspace imports for local, relative imports for published
- Eliminates @responsible-vibe/core resolution issues in published packages
- MCP inspector and CLI functionality both work with npx responsible-vibe-mcp
- Clean up old bundled files and entry package approach
