+++ package.json
@@ -3,9 +3,9 @@
   "version": "3.1.6",
   "description": "A Model Context Protocol server that acts as an intelligent conversation state manager and development guide for LLMs, featuring comprehensive long-term memory with persistent project artifacts",
   "type": "module",
-  "main": "index.js",
+  "main": "packages/cli/dist/index.js",
   "bin": {
-    "responsible-vibe-mcp": "index.js"
+    "responsible-vibe-mcp": "packages/cli/dist/index.js"
   },
   "engines": {
     "node": ">=18.0.0",
@@ -14,6 +14,7 @@
   "packageManager": "pnpm@9.14.2",
   "files": [
     "packages/*/dist/**/*",
+    "packages/*/package.json",
     "packages/docs/.vitepress/dist/**/*",
     "resources/**/*",
     "README.md",
@@ -50,6 +51,7 @@
     "@modelcontextprotocol/sdk": "1.17.5",
     "@types/js-yaml": "4.0.9",
     "js-yaml": "4.1.0",
+    "responsible-vibe-mcp": "file:responsible-vibe-mcp-3.1.6.tgz",
     "sqlite3": "^5.1.7",
     "zod": "^3.22.4"
   },
