+++ packages/mcp-server/src/tool-handlers/get-tool-info.ts
@@ -207,7 +207,7 @@ export class GetToolInfoHandler extends BaseToolHandler<
     // Build the complete response
     const response: GetToolInfoResponse = {
       tool_name: 'Responsible Vibe MCP - Development Workflow Management',
-      version: '3.1.6', // This should ideally come from package.json
+      version: '3.1.6-monorepo', // This should ideally come from package.json
       purpose:
         'Structured development workflows with guided phase transitions and conversation state management',
       description:
